refactor: GraphQL 관련 코드 정리
1. 미사용 common_app, auth_app, persona_app 삭제
2. graphQL 관련 코드를 model_types, mutations, queries, utils로 분리
3. URL 뒤의 trailing slash를 일괄 제거
