--- back/common_app/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class CommonAppConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'common_app'
