--- back/common_app/models.py
@@ -1,1 +1,0 @@
-from django.db import models
