+++ back/graphql_app/types/__init__.py
@@ -6,7 +6,7 @@ from .post import Mutation as PostMutation
 from .post import Query as PostQuery
 from .user import Mutation as UserMutation
 
-queries = (PostQuery, )
+queries = (PostQuery,)
 mutations = (UserMutation, PostMutation)
 
 Query = merge_types("Query", queries)
