+++ back/graphql_app/types/errors.py
@@ -1,4 +1,3 @@
-from typing import Optional
 import strawberry
