+++ back/graphql_app/types/post/__init__.py
@@ -1,0 +1,2 @@
+from .mutations import Mutation
+from .queries import Query
