+++ back/graphql_app/types/user/__init__.py
@@ -1,0 +1,1 @@
+from .mutations import Mutation
