+++ back/graphql_app/types/utils.py
@@ -5,7 +5,7 @@ def requires_auth(resolver):
     @wraps(resolver)
     def auth_wrapper(*args, **kwargs):
         if not kwargs['info'].context.request.user.is_authenticated:
-            raise AttributeError("auth required")
+            raise AttributeError("user required")
         return resolver(args, kwargs)
 
     return auth_wrapper
