--- back/persona_app/admin.py
@@ -1,5 +1,0 @@
-from django.contrib import admin
-
-from .models import Persona
-
-admin.site.register(Persona)
